+++ agent_chat/readme.md
@@ -1,6 +1,6 @@
 # Restack Agent Chat
 
-A sample repository with a an agent chat workflow.
+A sample repository with a an agent chat.
 
 ## Requirements
 
@@ -22,31 +22,31 @@ npm run dev
 ```
 
 This will start a Node.js app with Restack Services.
-Your code will be running and syncing with Restack to execute workflows or functions.
+Your code will be running and syncing with Restack to execute agents.
 
-## Run workflows
+## Run agents
 
 ### from UI
 
-You can run workflows from the UI by clicking the "Run" button.
+You can run agents from the UI by clicking the "Run" button.
 
-![Run workflows from UI](./chat_post.png)
+![Run agents from UI](./chat_post.png)
 
 ### from API
 
-You can run workflows from the API by using the generated endpoint:
+You can run agents from the API by using the generated endpoint:
 
-`POST http://localhost:6233/api/workflows/agentChat`
+`POST http://localhost:6233/api/agents/agentChat`
 
 ### from any client
 
-You can run workflows with any client connected to Restack, for example:
+You can run agents with any client connected to Restack, for example:
 
 ```bash
-npm schedule-workflow
+npm schedule-agent
 ```
 
-executes `scheduleWorkflow.ts` which will connect to Restack and execute the `agentChat` workflow.
+executes `scheduleAgent.ts` which will connect to Restack and execute the `agentChat` agent.
 
 ## Send events to the Agent
 
@@ -64,7 +64,7 @@ And see the events in the run:
 
 You can send events to the agent by using the following endpoint:
 
-`PUT http://localhost:6233/api/workflows/AgentWorkflow/:workflowId/:runId`
+`PUT http://localhost:6233/api/agents/AgentChat/:agentId/:runId`
 
 with the payload:
 
@@ -89,12 +89,12 @@ to end the conversation with the agent.
 
 ### from any client
 
-You can send event to the agent workflows with any client connected to Restack, for example:
+You can send event to the agent with any client connected to Restack, for example:
 
-Modify workflow_id and run_id in event_workflow.py and then run:
+Modify agentId and runId in eventAgent.ts and then run:
 
 ```bash
-pnpm event-workflow
+pnpm event-agent
 ```
 
 It will connect to Restack and send 2 events to the agent, one to generate another agent and another one to end the conversation.
